requirements.txt
streamlit==1.32.0
pandas==2.2.1
plotly==5.19.0
python-dotenv==1.0.1
langchain==0.1.13
langchain-google-genai==1.0.1
langchain-community==0.0.29
duckduckgo-search==5.2.1
faiss-cpu==1.8.0
SQLAlchemy==2.0.29
duckdb==0.10.1
